LayoutTests/fast/dom/Window/getComputedStyle-missing-parameter-expected.txt

Test that the first parameter to Window.getComputedStyle() is mandatory and not nullable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle() threw exception TypeError: Not enough arguments.
PASS window.getComputedStyle(null) threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
